app/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile app/requirements.in
blinker==1.9.0
    # via flask
click==8.1.8
    # via flask
flask==3.1.0
    # via -r app/requirements.in
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
numpy==2.2.5
    # via pandas
pandas==2.2.3
    # via -r app/requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.0
    # via -r app/requirements.in
pytz==2025.2
    # via pandas
redis==6.0.0
    # via -r app/requirements.in
six==1.17.0
    # via python-dateutil
tzdata==2025.2
    # via pandas
werkzeug==3.1.3
    # via flask
